Move classes to reflect package names; rename Model.scala files that were being ignored by model.* gitignore rule (and didn't contain a Model class anyway)
