+++ dl/src/main/scala/com/intel/analytics/bigdl/models/rnn/SimpleRNN.scala
