+++ dl/src/main/scala/com/intel/analytics/bigdl/models/vgg/VggForCifar10.scala
